better fix for micros/millis

+++ cores/portduino/linux/millis.cpp
@@ -6,18 +6,18 @@
 #include <stdio.h>
 #include <stdint.h>
 
-extern "C" uint32_t micros(void) {
+extern "C" unsigned long micros(void) {
   struct timeval te;
   gettimeofday(&te, NULL);                                  // get current time
-  uint32_t usecs = te.tv_sec * 1000000LL + te.tv_usec; // calculate
+  unsigned long usecs = te.tv_sec * 1000000LL + te.tv_usec; // calculate
   // printf("usecs %lu\n", usecs);
   return usecs;
 }
 
-extern "C" uint32_t millis(void) {
+extern "C" unsigned long millis(void) {
   struct timeval te;
   gettimeofday(&te, NULL); // get current time
-  uint32_t milliseconds =
+  unsigned long milliseconds =
       te.tv_sec * 1000LL + te.tv_usec / 1000; // calculate milliseconds
   // printf("milliseconds: %llu\n", milliseconds);
   return milliseconds;
